Declutter the build system a bit.
In src/sn_data.h test for the standard __linux__ macro instead of LINUX,
this way there is no need to arrange for it to be defined.  While at it,
add a guard for NETDEV_NR to fail with a meaningful message if none of
the OS-specific blocks have worked.

In src/Makefile.am remove AM_CFLAGS because the libraries belong to
Autoconf rather than Automake solution space; remove AM_CPPFLAGS because
all the the macros it defines are responsibility of Autoconf and/or not
used.  An example of the problem this fixes is that on an OS without
POSIX shared memory support Autoconf did not define HAVE_SHMGET in
config.h, but AM_CPPFLAGS defined it and the build failed because the
source code uses the conditional.

In src/sn_config.h include config.h so it actually takes Autoconf
results as intended, uncomment the check for HAVE_NCURSES_H because now
that works as intended.

In configure.ac remove checks that have no effect on the build, simplify
the usage of AC_CHECK_LIB() and AC_CHECK_HEADERS(), replace improvised
error messages with AC_MSG_ERROR(), make ncurses an optional dependency
because the code has conditionals to deal with it, leave a comment about
a potential future improvement.